requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
black==23.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
fastapi==0.93.0
frozenlist==1.3.3
h11==0.14.0
idna==3.4
motor==3.1.2
multidict==6.0.4
mypy-extensions==1.0.0
openai==0.27.1
packaging==23.1
pathspec==0.11.1
platformdirs==3.8.0
pydantic==1.10.5
pymongo==4.3.3
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.28.2
sniffio==1.3.0
starlette==0.25.0
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.6.3
urllib3==1.26.14
uvicorn==0.20.0
wincertstore==0.2
yarl==1.8.2
